更新代码

+++ src/common/HttpClient.jsx
@@ -1,6 +1,6 @@
 import $ from 'jquery';
 import {message} from 'antd';
-import {getQueryString} from './SystemFunction';
+import moment from 'moment';
 
 export const HttpClient = (function () {
     const REQUEST = "truth";
@@ -10,42 +10,55 @@ export const HttpClient = (function () {
     const requestServiceError = 2;
 
     //获取请求host
-    let httpClientHost = '';
+    let baseUrl = '';
 
     switch (process.env.NODE_ENV) {
         case "development":
-            httpClientHost = 'http://test01.triplego.cn:9527'; //测试主机
+            baseUrl = 'http://test01.triplego.cn/api'; //测试主机
+            // baseUrl = 'https://tripark.triplego.cn/api';
             break;
         case "production":
-            // httpClientHost = 'https://tripark.triplego.cn/api';
-            httpClientHost = 'http://test01.triplego.cn:9527'; //测试主机
+            // baseUrl = 'http://yingshi.triplego.cn/api';
+            baseUrl = 'http://test01.triplego.cn/api'; //测试主机
             break;
         default:
             break;
     }
 
     function query(url, rtype, data, callback, contentType = 'application/json;charset=UTF-8', processData = true) {
-        if (process.env.NODE_ENV === 'development') {
-            // console.log(location.hash.split('?')[1]);
-            if (location.hash.split('?')[1] && getQueryString(location.hash.split('?')[1], 'isMock') === 'true') {
-                httpClientHost = 'https://www.easy-mock.com/mock/5cd0f2f3682f200251f31dd3/immidiot';
-            } else if (location.hash.split('?')[1] && getQueryString(location.hash.split('?')[1], 'isDebug') === 'true') {
-                httpClientHost = '127.123.25:9090';
-            }
-        }
         let headers = null;
         let header_token = null;
-        if ((url === window.MODULE_PARKING_AUTHORITY + '/admin/token') || (url === window.MODULE_PARKING_AUTHORITY + '/configureInfo/getLogoConfig')) { // 判断是否是登录接口
+        const isLogin = (url.indexOf(window.MODULE_PARKING_AUTHORITY + '/admin/token') > -1) ||
+            (url.indexOf(window.MODULE_PARKING_AUTHORITY + '/configureInfo/getLogoConfig') > -1);
+        if (isLogin) { // 判断是否是登录接口
             header_token = "Basic Y29uc29sZTpjb25zb2xl";
-        } else {
+        }else {
             let access_token = window.customCookie.get('access_token') || '';
             header_token = "Bearer " + access_token;
         }
         headers = { "Authorization": header_token };
-        // console.log(REQUEST + "--" + httpClientHost + url + "--Params:", data);
+        if ((url.indexOf(window.MODULE_PARKING_AUTHORITY + '/admin/address/getVerificationCode') > -1)) {
+            headers = null
+        }
+        let httpUrl = '';
+        if (url.indexOf('http') > -1) {
+            httpUrl = url;
+            // 中间件接口，增加几个公共参数
+            if (url.indexOf('https://iotdev.triplego.cn/') > -1) {
+                if (!data.service) {
+                    data.service = url.substring((url.indexOf('cn/') + 3)).replace(/\//g, '.').replace(/(.\d+)/g, '');
+                }
+                data.version = '2.0';
+                data.timestamp = +moment();
+            }
+        } else if (url.indexOf('easy-mock') > -1) {
+            httpUrl = url.replace(/easy-mock/, 'https://www.easy-mock.com/mock/5cd0f2f3682f200251f31dd3/immidiot');
+        } else {
+            httpUrl = baseUrl + url
+        }
         $(function () {
             $.ajax({
-                "url": httpClientHost + url,
+                "url": httpUrl,
                 "async": true,
                 "cache": false,
                 "method": rtype,
@@ -60,38 +73,51 @@ export const HttpClient = (function () {
                 timeout: 40000,
                 "crossDomain": true,
                 success: function (d) {
-                    if (d.success) {
-                        callback(d, requestSuccess);
-                    } else {
-                        console.error('success=false: ', url);
-                        //可以处理d.error.code
-                        if (d.error.code === 10014 || d.error.code === 10015) {// 登录失效 || 用户不存在
-                            window.isInvalidToLogin = true;
-                            sessionStorage.clear();
-                            localStorage.clear();
-                            window.setPageMenu([]);
-                            window.setPermission({});
-                            window.setManagePartnerList([]);
-                            window.currentIsSystemAdmin = false;
-                            if (d.error.code === 10014) {
-                                location.hash = '/Login';
-                            } else {
-                                callback(d, requestSuccess);
-                                const hash = location.hash;
-                                if (!hash.match('Login')) {
-                                    location.hash = '/Login';
-                                }
-                            }
+                    if (url.indexOf('iotdev') > -1) {
+                        if (d.ack_code === 'ok') {
+                            callback(d, requestSuccess)
                         } else {
-                            message.error(d.error.message);
+                            message.error(d.err_msg);
                             callback(d, requestDataError);
+                        }
+                    } else {
+                        if (d.success) {
+                            callback(d, requestSuccess);
+                        } else {
+                            console.error('success=false: ', url);
+                            //可以处理d.error.code
+                            if (d.error.code === 10014 || d.error.code === 10015) {// 登录失效 || 用户不存在
+                                window.isInvalidToLogin = true;
+                                sessionStorage.clear();
+                                localStorage.clear();
+                                window.setPageMenu([]);
+                                window.setPermission({});
+                                window.setManagePartnerList([]);
+                                window.currentIsSystemAdmin = false;
+                                if (d.error.code === 10014) {
+                                    location.hash = '/Login';
+                                } else {
+                                    callback(d, requestSuccess);
+                                    const hash = location.hash;
+                                    if (!hash.match('Login')) {
+                                        location.hash = '/Login';
+                                    }
+                                }
+                            } else {
+                                message.error(d.error.message);
+                                callback(d, requestDataError);
+                            }
                         }
                     }
                 },
                 error: function (e) {
                     //服务异常
                     console.error(REQUEST + "--" + url + "--Error:", e);
-                    message.error("服务器异常！");
+                    if (e.status === 500) {
+                        message.error(e.responseJSON.message)
+                    } else {
+                        message.error("服务器异常！");
+                    }
                     callback(e, requestServiceError);
                 }
             });
@@ -99,7 +125,7 @@ export const HttpClient = (function () {
     }
 
     return {
-        ClientHost: httpClientHost,
+        ClientHost: baseUrl,
         GET: 'GET',
         POST: "POST",
         PUT: "PUT",
